[BUGFIX] Return always an integer in ProcessedFile::getSize
ProcessedFile::getSize() can call AbstractFile::getSize which can return
null instead of integer which is required as return type. Add a type
cast to avoid exceptions.

Backport the changes done already in v13 with #101327.

Resolves: #103777
Related: #101327
Releases: 12.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84445
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>

+++ typo3/sysext/core/Classes/Resource/ProcessedFile.php
@@ -507,9 +507,9 @@ class ProcessedFile extends AbstractFile
     public function getSize(): int
     {
         if ($this->usesOriginalFile()) {
-            return $this->getOriginalFile()->getSize();
+            return (int)$this->getOriginalFile()->getSize();
         }
-        return parent::getSize();
+        return (int)parent::getSize();
     }
 
     /**
